moves reporting outside of celery
The previous model of reading Celery results was likely the correct way
to generate a report, but a good portion of completed work was left
un-reporting as the main thread was busy parsing sitemaps. This commit
provides a temporary fix that results are persisted to disk earlier
(sitemap parsing will continue throughout via workers).